NotFoundException, validatiors, frontend changes

+++ src/SimpleAgileBoard.Application/Lanes/Commands/DeleteLane/DeleteLaneCommandHandler.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using MediatR;
 using SimpleAgileBoard.Application.Boards.Queries.GetBoard;
 using SimpleAgileBoard.Application.Boards.Services;
+using SimpleAgileBoard.Application.Common.Exceptions;
 using SimpleAgileBoard.Application.Lanes.Services;
 
 namespace SimpleAgileBoard.Application.Lanes.Commands.DeleteLane
@@ -22,10 +23,18 @@ namespace SimpleAgileBoard.Application.Lanes.Commands.DeleteLane
         public async Task<BoardViewModel> Handle(DeleteLaneCommand request, CancellationToken cancellationToken)
         {
             var lane = await _laneRepository.Get(request.LaneId, cancellationToken);
-
+            if (lane == null)
+            {
+                throw new NotFoundException(nameof(lane), request.LaneId);
+            }
+            
             await _laneRepository.Remove(lane, cancellationToken);
             var board = await _boardRepository.Get(request.BoardId, cancellationToken);
-
+            if (board == null)
+            {
+                throw new NotFoundException(nameof(board), request.BoardId);
+            }
+            
             return new BoardViewModel
             {
                 Board = board
